requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
# pymongo[srv]==3.6

##-e .